legacy-openai-impl
